neusoft-python-backend/requirements.txt
acvl_utils==0.2.5
anyio==4.12.0
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blinker==1.9.0
blosc2==3.12.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
connected-components-3d==3.26.1
contourpy==1.3.3
cycler==0.12.1
dicom2nifti==2.6.2
dynamic_network_architectures==0.4.2
einops==0.8.1
fft-conv-pytorch==1.2.0
filelock==3.20.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
imagecodecs==2025.11.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
monai==1.5.1
mpmath==1.3.0
msgpack==1.1.2
ndindex==1.10.1
networkx==3.6.1
nibabel==5.3.3
nnunetv2==2.6.2
numexpr==2.14.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
py-cpuinfo==9.0.0
pyarrow==22.0.0
pydicom==3.0.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-gdcm==3.2.2
pytorch-ignite==0.5.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
simpleitk==2.5.3
six==1.17.0
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
TotalSegmentator==2.12.0
tqdm==4.67.1
traceback2==1.4.0
trimesh==4.10.1
typer-slim==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
unittest2==1.1.0
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
xvfbwrapper==0.2.16
yacs==0.1.8
